Move metadata accessors to modelmeta module
